feat: refactor payment page components and improve session handling
